REST method for batch adding songs to playlist.

+++ kiara-server/src/main/java/uk/co/yojan/kiara/server/resources/PlaylistResource.java
@@ -25,57 +25,102 @@ import static uk.co.yojan.kiara.server.OfyService.ofy;
 public class PlaylistResource {
   private static Logger log = Logger.getLogger(Playlist.class.getName());
 
+  /*
+   * GET request handler for all playlists for a given user.
+   *
+   * @param userId  the user id (Spotify id) to get the playlists for
+   * @param detail  true if extra data about the playlist's songs is also to be sent
+   * @param request  the request object that triggered the handler
+   * @return if E-tag is valid, a 304 Not Modified otherwise the updated resource
+   */
   @GET
   @Produces(MediaType.APPLICATION_JSON)
   public Response getAll(@PathParam("user_id") String userId,
                          @DefaultValue("false") @QueryParam("detail") boolean detail,
                          @Context Request request) {
     User u = ofy().load().key(Key.create(User.class, userId)).now();
+
     EntityTag etag = new EntityTag(Long.toString(u.v()));
     Response.ResponseBuilder builder = request.evaluatePreconditions(etag);
     CacheControl cc = new CacheControl();
-//    cc.setMaxAge(60);
+
     if(builder != null) {
       Logger.getLogger("PlaylistResource").info("Cached resource did change " + etag.getValue());
       return builder.build();
     }
+
     // cached resource did change, serve updated content.
-//    if(builder == null) {
-      if(detail) {
-        builder = Response.ok().entity(u.getPlaylistsWithSongs()).tag(etag);
-      } else {
-        builder = Response.ok().entity(u.getAllPlaylists()).tag(etag);
-      }
-//    }
+    if(detail) {
+      builder = Response.ok().entity(u.getPlaylistsWithSongs()).tag(etag);
+    } else {
+      builder = Response.ok().entity(u.getAllPlaylists()).tag(etag);
+    }
     return builder.cacheControl(cc).build();
   }
 
+  /*
+   * GET request handler for a specific playlist for the user.
+   *
+   * Fetches the user, followed by the playlist and if the detail param is
+   * set to true, the songs from the Datastore.
+   *
+   * @param userId  the user id (Spotify id) to get the playlist for
+   * @param id  the playlist identifier (long)
+   * @param detail  whether to also send the playlist's song data.
+   * @return if E-tag is valid, 304 Not Modified otherwise the updated resource.
+   */
   @GET
   @Produces(MediaType.APPLICATION_JSON)
   @Path("/{id}")
   public Response get(@PathParam("user_id") String userId,
                       @PathParam("id") Long id,
-                      @DefaultValue("false") @QueryParam("detail") boolean detail) {
+                      @DefaultValue("false") @QueryParam("detail") boolean detail,
+                      @Context Request request) {
     User u = ofy().load().key(Key.create(User.class, userId)).now();
-    if(detail) {
-      Playlist playlist = u.getPlaylist(id);
-      ArrayList<Song> songs = new ArrayList<Song>(playlist.getAllSongs());
-      return Response.ok().entity(new PlaylistWithSongs(playlist, songs)).build();
+    Playlist playlist = u.getPlaylist(id);
+
+    EntityTag etag = new EntityTag(Long.toString(playlist.v()));
+    Response.ResponseBuilder builder = request.evaluatePreconditions(etag);
+    CacheControl cc = new CacheControl();
+
+    if(builder != null) {
+      Logger.getLogger("PlaylistResource").info("Cached playlist resource did not change " + etag.getValue());
+      return builder.build();
     }
 
-    Playlist p = u.getPlaylist(id);
-    return Response.ok().entity(p).build();
+    if(detail) {
+      ArrayList<Song> songs = new ArrayList<Song>(playlist.getAllSongs());
+      builder = Response.ok().entity(new PlaylistWithSongs(playlist, songs));
+    } else {
+      builder =  Response.ok().entity(playlist);
+    }
+
+    return builder.cacheControl(cc).build();
   }
 
+  /*
+   * DELETE the playlist resource associated with the param id.
+   *
+   * @param userId  the user id whose playlist to delete.
+   * @param id  the playlist identifier to delete.
+   * @return Not Modified if playlist was not deleted, else 200 OK.
+   */
   @DELETE
   @Path("/{id}")
   public Response delete(@PathParam("user_id") String userId, @PathParam("id") Long id) {
     User owner = ofy().load().key(Key.create(User.class, userId)).now();
-    owner.removePlaylist(id);
-    ofy().delete().key(Key.create(Playlist.class, id)); // async
-    return Response.ok().build();
+
+    if(owner.removePlaylist(id)) {
+      ofy().delete().key(Key.create(Playlist.class, id)); // async
+      return Response.ok().build();
+    } else {
+      return Response.notModified().build();
+    }
   }
 
+  /*
+   * PUT (not used yet) handler to update the resource.
+   */
   @PUT
   @Consumes(MediaType.APPLICATION_JSON)
   @Produces(MediaType.APPLICATION_JSON)
@@ -92,6 +137,13 @@ public class PlaylistResource {
     }
   }
 
+  /*
+   * POST handler to create a new playlist resource.
+   *
+   * @param userId  the user id to who the playlist belongs.
+   * @param uri  the uri from the request context.
+   * @param item  the JSON playlist resource to create.
+   */
   @POST
   @Consumes(MediaType.APPLICATION_JSON)
   @Produces(MediaType.APPLICATION_JSON)
@@ -99,7 +151,6 @@ public class PlaylistResource {
                          @Context UriInfo uri,
                          Playlist item,
                          @Context Request request) {
-//    return Response.ok().build();
     Result<User> ownerResult = ofy().load().key(Key.create(User.class, userId)); // async
     item.updateLastViewed();
     ofy().save().entity(item).now(); // sync
